Move loop and heartline roll TEDs to their own file
